pyproject.toml
[project]
name = "graph-rag-recipes"
version = "0.1.0"
description = "GraphRAG 菜谱智能推荐系统"
readme = "README.md"
authors = [
    { name = "lyzno1", email = "yuanyouhuilyz@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.2",
    "pandas>=2.1",
    "numpy>=1.26",
    "requests>=2.31",
    "scikit-learn>=1.4",
    "sentence-transformers>=3.0",
    "streamlit>=1.32",
    "openai>=1.35",
    "httpx[socks]>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
graph-rag-recipes = "graph_rag_recipes:main"

[project.optional-dependencies]
dev = ["pytest>=8.2"]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
